requirements.txt
fastapi[all]
sqlmodel
asyncer
pydantic-sqlalchemy
rich
fastapi-template
fastapi-login
python-jose
databases